feat(preprocessor): Implement UCN and wide char literal support
- Update `PPTokenKind::CharLiteral` and `TokenKind::CharacterConstant` to store `u64` instead of `u8`.
- Update `Literal::Char` to store `u64`.
- Implement `parse_ucn_escape` in `src/ast/literal_parsing.rs` to handle `\u` and `\U` escape sequences.
- Update `PPLexer::lex_char_literal` to use `literal_parsing::parse_char_literal` and `u64` type.
- Update `AstDumper` to format `LiteralChar` using unicode characters.
- Add regression test `src/tests/pp_wide_char.rs`.
- Verify all tests pass.

This enables correct evaluation of wide character constants (e.g. `L'\u0400'`) in preprocessor directives and supports UCNs in character literals.

+++ src/parser/lexer.rs
@@ -11,7 +11,7 @@ pub enum TokenKind {
     // === LITERALS ===
     IntegerConstant(i64, Option<IntegerSuffix>), // Parsed integer literal value
     FloatConstant(f64, Option<FloatSuffix>),     // Parsed float literal value
-    CharacterConstant(u8),                       // Byte value of character constant
+    CharacterConstant(u64),                      // Value of character constant
     StringLiteral(StringId),                     // Interned string literal
 
     // === IDENTIFIERS ===
